Refactor method name and variables for improved clarity
Changed the name of the method "openCloseFind" to "findMatchingPair" in "string_view.hpp" to better reflect its functionality. It also improves consistency with other similar methods. The internal variable "passed_pairs" was changed to "pairs_count" and "it_to_elem" was changed to "matched_pair_iterator" to enhance code readability. This refactor does not affect the method's functionality.