"Remove dependency on range-v3 library"
This commit removed the range-v3 library from the project and modified functions depending on it.

The use of filter from the range-v3 was replaced with regular for-loops and `if` conditional checks. The removal reduces dependencies, simplifies the code and also improves readability. Since range-v3 was no longer needed, it was removed from vcpkg.json.